Don't bother using 'half' type in HLSL as it's mapped to float anyway. Oops.

+++ src/d2dx/DisplayIntegerScalePS.hlsl
@@ -20,7 +20,7 @@
 
 Texture2D sceneTexture : register(t0);
 
-half4 main(
+float4 main(
     in noperspective float2 tc : TEXCOORD0) : SV_TARGET
 {
 	return sceneTexture.SampleLevel(PointSampler, tc, 0);
